...G...PS. [ZBX-12983] fixed compilation warnings under Windows
git-svn-id: svn://svn.zabbix.com/trunk@74284 97f52cf1-0a1b-0410-bd0e-c28be96e8082

+++ src/libs/zbxwin32/fatal.c
@@ -146,7 +146,7 @@ static void	print_backtrace(CONTEXT *pctx)
 	if (0 != GetModuleFileNameEx(hProcess, NULL, szProcessName, ARRSIZE(szProcessName)))
 	{
 		char	*ptr;
-		int	path_alloc = 0, path_offset = 0;
+		size_t	path_alloc = 0, path_offset = 0;
 
 		process_name = zbx_unicode_to_utf8(szProcessName);
